Adding Word and Phrase-Level Conditioning with both categorical and legendre parameters.
This model expects input tensors in the form of labels per input
symbols for both categorical and continuous conditioning. The paths
of which be specified in the metadata input file under the correct
column heading (see fastpitch/data_function.py for correct heading)

I still have to add a parameter for the legendre coefficients since
the default is currently hard-coded to 3 for word-level in model.py.

Still need to document this code in the .md file.